Start decoupling miner and store loading from conf
