+++ collector/stores/stores.go
@@ -1,0 +1,13 @@
+package miners
+
+const (
+	// Store type constants
+	INFLUXDB2_TYPE string = "influxdb2"
+)
+
+type Store interface {
+	// SendCGMinerMetrics(miner *Miner, metrics *metrics.CGMinerMetrics) error
+	/** TODO:
+	SendBitaxeMetrics(...)
+	*/
+}
